# yorindo-api/openapi.yaml
openapi: 3.0.3
info:
  title: Yorindo API
  version: 1.0.0
  license:
    name: Proprietary
    url: https://yorindo.id/license
  description: |
    Yorindo event management platform API.

    All dates are ISO 8601 UTC strings (e.g., "2026-04-15T02:00:00.000Z").
    All IDs are opaque strings.
    All errors use { error: { code, message, details[] } } format exclusively.
    Rate limit: POST /api/registrations is limited to 10 requests/hour per IP.
servers:
  - url: http://localhost:3000/api
    description: Local development (via Docker Compose)
  - url: https://api.yorindo.id/api
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              description: SCREAMING_SNAKE_CASE error code
              example: VALIDATION_ERROR
            message:
              type: string
              description: Human-readable error description
              example: Validation failed
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: email
                  message:
                    type: string
                    example: Invalid email format

    # ─── Async Job ─────────────────────────────────────────────────────────────
    AsyncJobResponse:
      type: object
      required: [jobId, status]
      properties:
        jobId:
          type: string
          example: etl-upload-001
        status:
          type: string
          enum: [queued]
          example: queued

    # ─── Lookup Tables ─────────────────────────────────────────────────────────
...
    DataRightsResponse:
      type: object
      required: [message, requestId]
      properties:
        message:
          type: string
          example: Request queued
        requestId:
          type: string
          example: data-rights-request-001
paths:
...
  /etl/upload:
    post:
      summary: Trigger ETL contact import job (admin only)
      operationId: triggerEtlUpload
      tags: [ETL]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: ETL job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncJobResponse'
        '400':
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /etl/jobs/{jobId}:
    get:
      summary: Get ETL job status
      operationId: getEtlJobStatus
      tags: [ETL]
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ETL job status payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /participants/erasure-request:
    post:
      summary: Request participant data erasure
      operationId: requestParticipantErasure
      tags: [Participants]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, email]
              properties:
                phone:
                  type: string
                  minLength: 8
                email:
                  type: string
                  format: email
      responses:
        '202':
          description: Erasure request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataRightsResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Suppression already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # ─────────────────────────────────────────────────────────────────────────────
  # SMART FILTER
  # ─────────────────────────────────────────────────────────────────────────────
...