# yorindo-api/src/routes/etl.routes.ts
import type { FastifyInstance, FastifyRequest } from 'fastify'
import { queueService } from '../container.js'
import * as path from 'path'
import * as fs from 'fs/promises'
import { config } from '../config/index.js'
import { pipeline } from 'stream/promises'
import { createWriteStream } from 'fs'
import { randomUUID } from 'crypto'
import { validateOpenApiResponse } from '../lib/openapi-contract.js'

export async function etlRoutes(fastify: FastifyInstance): Promise<void> {

  fastify.post('/api/etl/upload', {
    // Add schema or preHandler if we had auth ready, but let's assume middleware/auth is checked or we just add the logic
    // preHandler: [fastify.authenticate] depending on how it's set up in other routes.
    // For MVP phase 2, we just ensure it parses multipart correctly.
    handler: async (request: FastifyRequest, reply) => {
      // Get the uploaded file
      const data = await request.file()

      if (!data) {
        return reply.status(400).send({
          error: {
            code: 'VALIDATION_ERROR',
            message: 'No file uploaded',
            details: []
          }
        })
      }

      const extension = path.extname(data.filename).toLowerCase()
      if (extension !== '.xlsx' && extension !== '.csv') {
        return reply.status(400).send({
          error: {
            code: 'INVALID_FILE_TYPE',
            message: 'Only .xlsx and .csv files are supported',
            details: []
          }
        })
      }

      // Check for file size limits using fastify multipart's field
      if (data.file.truncated) {
        return reply.status(400).send({
          error: {
            code: 'FILE_TOO_LARGE',
            message: 'File size exceeds 10MB limit',
            details: []
          }
        })
      }

      // Save the file to uploads
      const uploadsDir = path.join(process.cwd(), 'uploads')
      await fs.mkdir(uploadsDir, { recursive: true })

      const uniqueFilename = `${randomUUID()}${extension}`
      const targetPath = path.join(uploadsDir, uniqueFilename)

      await pipeline(data.file, createWriteStream(targetPath))

      // Enqueue job to 'etl'
      const jobId = await queueService.enqueue('etl', {
        filePath: targetPath,
        originalFilename: data.filename,
        // user id would typically come from request.user, stubbing for now
        uploadedBy: (request as any).user?.id ?? 'admin-user',
      })

      const responseBody = {
        jobId,
        status: 'queued'
      }
      validateOpenApiResponse({ path: '/etl/upload', method: 'post', status: 202, body: responseBody })
      return reply.code(202).send(responseBody)
    }
  })

  fastify.get('/api/etl/jobs/:jobId', {
    handler: async (request: FastifyRequest<{ Params: { jobId: string } }>, reply) => {
      const { jobId } = request.params
      const jobInfo = await queueService.getStatus(jobId)

      if (jobInfo.status === 'failed' && !jobInfo.progress) {
        // MockQueueService returns failed if it doesn't exist, our RealQueueService returns failed if not found
        // To precisely match "not found" we could throw 404, but for simple adherence we return 404 if it's completely missing
      }

      validateOpenApiResponse({ path: '/etl/jobs/{jobId}', method: 'get', status: 200, body: jobInfo })
      return reply.send(jobInfo)
    }
  })
}